+++ ui/text.h
@@ -18,7 +18,7 @@ struct Font
 
 extern Font FONT_DEFAULT;
 
-glm::ivec2 drawText(string text, glm::ivec2 position,
+glm::ivec2 drawText(string text, glm::vec2 position,
                     const Font *font = &FONT_DEFAULT);
 
 } // namespace
